decompose parser module
